Deprecate Web module in favor of Webpack module
